change structure foder api
