+++ src/service/index.js
@@ -1,0 +1,2 @@
+export * from './tweb3';
+export * from './api';
